Исправлены замечания по ДЗ
