--- Navigation/Coordinators/CoordinatorProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  CoordinatorProtocol.swift
-//  Navigation
-//
-//  Created by Филипп Степанов on 02.05.2022.
-//
-
-import Foundation
-
-protocol Coordinator {
-    func start()
-}
